Renamed keys to positions (which they are)

+++ h2o-core/src/main/java/water/udf/TypedChunk.java
@@ -7,8 +7,8 @@ import water.fvec.Vec;
  * Represents a chunk that knows its type
  */
 public interface TypedChunk<T> extends Vec.Holder {
-  T get(int i);
-  boolean isNA(int i);
+  T get(long i);
+  boolean isNA(long i);
   int length();
   long start();
   Chunk rawChunk();
